+++ source/Options/styles.scss
@@ -1,8 +1,6 @@
 @import "../styles/fonts";
 @import "../styles/variables";
 
-@import "~webext-base-css/webext-base.css";
-
 body {
   color: $black;
   background-color: $greyWhite;
