+++ webpack.config.js
@@ -61,6 +61,7 @@ module.exports = {
     manifest: path.join(sourcePath, 'manifest.json'),
     background: path.join(sourcePath, 'Background', 'index.ts'),
     contentScript: path.join(sourcePath, 'ContentScript', 'index.ts'),
+    injected: path.join(sourcePath, 'Injected', 'index.ts'),
     popup: path.join(sourcePath, 'Popup', 'index.tsx'),
     options: path.join(sourcePath, 'Options', 'index.tsx'),
     visualization: path.join(sourcePath, 'Visualization', 'index.tsx'),
